Empty deployer module
